Support podman in macOS
With [a fix][1] in [podman v3.4.3][2], these commands now work as
expected in macOS. Potentially, it might make sense to version check
podman to ensure that the minimum version is met, but I'm not sure
that's needed because it's unlikely that people have an older version
installed _and_ wish to use this tool.

I'm unsure whether the commands work on Windows so I left the
unsupported version there compiling with the negation of the supported
flags.

[1]: https://github.com/containers/podman/issues/12402
[2]: https://github.com/containers/podman/blob/4ba71f955a944790edda6e007e6d074009d437a7/RELEASE_NOTES.md#bugfixes-2