feat: improve "Fetching" message (wagoodman/dive#482)
The Fetching... message was confusing.

This replaces it with a clearer messages to avoid confusion.

Additional fix: show original error unless image is not found

Only try doing a pull if the image isn't found. Everything else should
just generate the error so the user can fix it.

Fixes wagoodman/dive#360

Co-authored-by: Christian Höltje <docwhat@gerf.org>